Added Ai, Action, Animation, GraphicsComponent

+++ TestGame/src/main/java/com/donishchenko/testgame/resources/ResourceLoader.java
@@ -12,7 +12,7 @@ import java.util.HashMap;
 
 public class ResourceLoader {
 
-    public static double SCALE = 0.5;
+    public static float SCALE = 0.5f;
     private static final HashMap<String, HashMap<String, Object>> assets = new HashMap<>();
 
     static {
